requirements.txt
numpy
pandas
python-dotenv
matplotlib
seaborn
plotly
scikit-learn
streamlit
nltk